News Article Sorting/requirements.txt
certifi==2021.10.8
click==8.0.4
colorama==0.4.4
Flask==2.0.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
joblib==1.1.0
MarkupSafe==2.1.1
nltk==3.7
numpy==1.21.5
pymongo==4.0.2
PyYAML==6.0
pyyml==0.0.2
regex==2022.3.15
scikit-learn==1.0.2
scipy==1.7.3
sklearn==0.0
threadpoolctl==3.1.0
tqdm==4.63.1
typing_extensions==4.1.1
Werkzeug==2.0.3
wincertstore==0.2
zipp==3.7.0
